requirements.txt
gpiozero
Pillow
rpi.gpio
picamera2
opencv-python
numpy
PyYAML